Exception checkups

+++ src/CgiHandler.cpp
@@ -1,10 +1,15 @@
 #include "../inc/CgiHandler.hpp"
+#include "../inc/Signal.hpp"
 #include <stdexcept>
 
 t_cgi_state	CgiHandler::checkCgi()
 {
 	int	status;
 
+	if (_state == CGI_FAILED)
+	{
+		return (CGI_FAILED);
+	}
 	if (std::time(nullptr) - _start_time >= _CGI_TIMEOUT_S){
 		std::cout << "KILL THE CHILDREN\n";
 		kill(_pid, SIGKILL);
@@ -17,6 +22,7 @@ t_cgi_state	CgiHandler::checkCgi()
 			}
 		}
 		std::cout << "CGI script exited succesfully" << std::endl;
+		
 		// std::cout << "\n\nCGI_READY, OUTPUT:\n" << std::endl;
 		//
 		// std::fstream fafa(output_filename);
@@ -31,7 +37,6 @@ t_cgi_state	CgiHandler::checkCgi()
 	return (CGI_WAITING);
 }
 
-// make a getter to CGI's type (Python) {}
 
 void	getCgiEnv(Request &request, const ServerConfig &config, std::vector<char*> &envp, std::vector<std::string> &envVars)
 {
@@ -45,8 +50,8 @@ void	getCgiEnv(Request &request, const ServerConfig &config, std::vector<char*> 
 	std::string pathInfo = "PATH_INFO=";
 	std::string scriptName = "SCRIPT_NAME=" + request.getLocation().root;
 	std::string	serverPort = "SERVER_PORT=" + std::to_string(config.listen_port);
-	std::string serverName = "SERVER_NAME=" + config.server_names[0]; // or should we just put IP here? we might have multiple names I'm confutse
-	std::string remoteAddr = "REMOTE_ADDR=" + config.listen_ip; // "The IP address of remote host", should this really be us?
+	std::string serverName = "SERVER_NAME=" + config.server_names[0];
+	std::string remoteAddr = "REMOTE_ADDR=" + config.listen_ip;
 	std::string serverProtocol = "SERVER_PROTOCOL=HTTP/1.1";
 	std::string redirectStatus = "REDIRECT_STATUS=200";
 	std::string gatewayInterface = "GATEWAY_INTERFACE=CGI/1.1";
@@ -109,25 +114,13 @@ void	CgiHandler::launchCgi(Request &request)
 
 	int	output_fd = open(output_filename.c_str(), O_RDWR | O_CREAT | O_TRUNC, 0644); // let's verify permissions later
 	if (output_fd < 0){
-		// close(input_fd);
-		// throw
+		_state = CGI_FAILED;
+		throw std::runtime_error("Failed to open file in launchCgi");
 	}
-	
 	getCgiEnv(request, request.getConfig(), envp, envVars);
-	// we have saved the post body into infile before
-	// close(input_fd); // move file position indicator to the beginning of the file stream by closing
 
-	std::string	executable;
-	if (endsWith(request.getPath(), ".py")) {
-		executable = request.getLocation().cgi_path_py;
-	} else if (endsWith(request.getPath(), ".bla")) {
-		executable = request.getLocation().cgi_path_bla;
-	} else {
-		throw std::runtime_error("launchCgi(): request path ending unsupported");
-	}
+	std::string	executable = request.getLocation().cgi_path_py;
 	std::string	script = request.getLocation().root + request.getPath();
-
-	// std::cout << "\n\nCGI_PATH: " << executable << "\n\nrequest.path: " << request.getPath() << "\n\nSCRIPT_PATH: " << script << std::endl;
 
 	char *argv[] = {
 		(char*)executable.c_str(),
@@ -135,13 +128,11 @@ void	CgiHandler::launchCgi(Request &request)
 		nullptr
 	};
 
-	// std::cout << "\n\n\n\nCGIHANDLER\n\n\n\n";
 	int pid = fork();
 
 	if (pid < 0) {
-		// close(input_fd);
 		close(output_fd);
-		throw std::runtime_error("fork() failed"); // are we closing the server?
+		throw std::runtime_error("fork() failed");
 	}
 	else if (pid == 0) {
 		int input_fd = open(request.getPostBodyFilename().c_str(), O_RDONLY, 0644);
@@ -153,7 +144,7 @@ void	CgiHandler::launchCgi(Request &request)
 		dup2(output_fd, STDOUT_FILENO);
 		close(output_fd);
 		execve(argv[0], argv, envp.data());
-		// throw something ?
+		// throw std::runtime_error("CGI: execve failed");
 	}
 	else {
 		_start_time = std::time(nullptr);
